backend/requirements.txt
# Render-specific requirements with compatible versions
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.9.0
boto3==1.34.0
# Use psycopg2-binary for easier installation
psycopg2-binary==2.9.9
gunicorn==21.2.0
